backend/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-environ==0.4.5
django-cors-headers==3.7.0
djangorestframework-simplejwt